requirement.txt
spacy
transformeres
pymupdf
pytesseract
pillow
en_core_web_sm-3.8.0-py3-none-any.whl
numpy